update tests import and snowflake module name
