+++ src/snowflake_utils/__init__.py
@@ -1,0 +1,5 @@
+from .snowflake_utils import create_logger, get_connection, write_to_snowflake, table_exists
+
+logger = create_logger("snowflake")
+
+__all__ = ["logger", "get_connection", "write_to_snowflake", "table_exists"]
