+++ src/snowflake_utils/snowflake_utils.py
@@ -1,10 +1,76 @@
 import inspect
+import logging
+import os
 
 import pandas as pd
+import snowflake.connector
 from snowflake.connector.pandas_tools import write_pandas
 
-from snowflake import logger
 
+def create_logger(name: str = "snowflake") -> logging.Logger:
+    """Create or retrieve a module-level logger.
+
+    Args:
+        name (str): Logger name (default: 'snowflake')
+
+    Returns:
+        logging.Logger: Configured logger
+    """
+    logger = logging.getLogger(name)
+
+    # Only configure once (avoids duplicate handlers on import)
+    if not logger.handlers:
+        handler = logging.StreamHandler()
+        formatter = logging.Formatter("%(asctime)s [%(levelname)s] %(name)s: %(message)s")
+        handler.setFormatter(formatter)
+        logger.addHandler(handler)
+        logger.setLevel(logging.INFO)
+    
+    return logger
+
+
+logger = create_logger("snowflake_utils")
+
+
+def get_connection() -> snowflake.connector.SnowflakeConnection:
+    """Create and return a Snowflake connection using environment variables.
+
+    Expected environment variables:
+        SNOWFLAKE_USER
+        SNOWFLAKE_PASSWORD
+        SNOWFLAKE_ACCOUNT
+        SNOWFLAKE_DATABASE
+        SNOWFLAKE_SCHEMA
+        SNOWFLAKE_WAREHOUSE
+        SNOWFLAKE_ROLE
+    """
+    required_vars = [
+        "SNOWFLAKE_USER",
+        "SNOWFLAKE_PASSWORD",
+        "SNOWFLAKE_ACCOUNT",
+        "SNOWFLAKE_DATABASE",
+        "SNOWFLAKE_SCHEMA",
+        "SNOWFLAKE_WAREHOUSE",
+        "SNOWFLAKE_ROLE",
+    ]
+    missing = [v for v in required_vars if not os.getenv(v)]
+    if missing:
+        raise EnvironmentError(f"Missing required Snowflake env vars: {missing}")
+
+    logger.info("Connecting to Snowflake...")
+    conn = snowflake.connector.connect(
+        user=os.getenv("SNOWFLAKE_USER"),
+        password=os.getenv("SNOWFLAKE_PASSWORD"),
+        account=os.getenv("SNOWFLAKE_ACCOUNT"),
+        database=os.getenv("SNOWFLAKE_DATABASE"),
+        schema=os.getenv("SNOWFLAKE_SCHEMA"),
+        warehouse=os.getenv("SNOWFLAKE_WAREHOUSE"),
+        role=os.getenv("SNOWFLAKE_ROLE"),
+    )
+    logger.info("Connected to Snowflake successfully.")
+    return conn
+
+    
 def table_exists(conn, table_name: str, schema: str | None = None, database: str | None = None) -> bool:
     """Check if a Snowflake table exists."""
     schema = schema or conn.schema
